backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-filter==24.2
dj-database-url==2.1.0
psycopg[binary]
django-cors-headers==4.3.1
python-dotenv==1.0.1
google-generativeai==0.6.0